backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.2.1
redis==5.0.4
httpx==0.27.0
python-multipart==0.0.9
